Update Notification model
