--- src/main/java/com/example/eventplanner/model/user/Notification.java
@@ -1,25 +1,0 @@
-package com.example.eventplanner.model.user;
-
-import com.example.eventplanner.model.Entity;
-import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.SQLRestriction;
-
-import java.util.Date;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@SQLRestriction("active = true")
-@jakarta.persistence.Entity
-public class Notification extends Entity {
-    private String message;
-    private Date dateSent;
-    private boolean seen;
-    @ManyToOne
-    private BaseUser user;
-}
